requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pyodbc==5.1.0
pydantic==2.8.0
python-dotenv==1.0.1
